Remove wrapping; address overflows; fix #7
There were a couple issues causing the overflows. First, resume was
recursive but should have been a loop. This usually didn't cause a
problem before because the recursion depth was never high enough.
However, one of the coercions that bind called in turn called map,
and map called bind again, leading to an infinite loop. This only
occurred if Subs was used which is why the factorial example didn't
trigger it.

I also removed the struct wrapper but in hindsight I don't think that
caused any problems (aside from maybe efficiency). I will probably add
it back in the future.

So the lesson is to be extra careful about which functions are
called during the main loop. Some sort of call graph visualization
would have helped uncover this sooner.